Adjusted to pure jQuery
Pulled out the new element, since that seemed like overkill, and just
grabbed valid <blink> tags with jQuery. This was pulled straight from
Wikipedia, so nothing special, but makes more sense as an
implementation. We lose the properties, but we can come back and add
those in if we decide we want them.